Extract page helpers and clean modules
